requirements_minimal.txt
# UI
customtkinter==5.2.1
Pillow==10.1.0  # Required by customtkinter

# Core functionality
wikipedia-api==0.6.0
duckduckgo-search==3.9.3

# Utilities
python-dotenv==1.0.0  # For environment variables
